requirements_test.txt
-r requirements.txt
alabaster==0.7.6
Babel==2.0
binaryornot==0.3.0
click==3.3
cookiecutter==1.0.0
coverage==3.7.1
coveralls==0.5
docopt==0.6.2
docutils==0.12
Flask-WTF==0.12
funcsigs==0.4
mock==1.3.0
pbr==1.4.0
pluggy==0.3.0
py==1.4.30
Pygments==2.0.2
pytest==2.7.2
pytest-flask==0.8.1
PyYAML==3.11
requests==2.7.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
termcolor==1.1.0
tox==2.1.1
travis-solo==0.0.9
virtualenv==13.1.0
WTForms==2.0.2
pep8==1.6.2
flask-script==2.0.5
